pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mysql-mcp"
version = "0.1.0"
description = "A modern MySQL Model Context Protocol (MCP) server built with FastMCP"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Michael Zag", email = "Michael@MichaelZag.com" }
]
keywords = ["mcp", "mysql", "database", "ai", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.11.3",
    "mysql-connector-python>=9.4.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/mysql-mcp"
Repository = "https://github.com/yourusername/mysql-mcp"
Issues = "https://github.com/yourusername/mysql-mcp/issues"
Documentation = "https://github.com/yourusername/mysql-mcp#readme"

[project.scripts]
mysql-mcp = "mysql_mcp.server:main"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.10",
]

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.black]
line-length = 88
target-version = ['py313']

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.hatch.build.targets.wheel]
packages = ["src/mysql_mcp"]
